make: Spring CRUD tests

+++ SQLviaSpringDataJPA/src/main/java/com/sqlviaspringdatajpa/service/ProductService.java
@@ -2,34 +2,40 @@ package com.sqlviaspringdatajpa.service;
 
 import com.sqlviaspringdatajpa.model.Product;
 import com.sqlviaspringdatajpa.repository.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class ProductService {
 
     private final ProductRepository productRepository;
 
-    @Autowired
-    public ProductService(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
     public Product findById(Integer id) {
-        return productRepository.getById(id);
+        return productRepository.findById(id).orElse(new Product());
     }
 
     public List<Product> findAll() {
         return productRepository.findAll();
     }
 
+    public Product findByName(String name) {
+        return productRepository.existsProductByProductName(name) ?
+                productRepository.findProductByProductName(name) :
+                new Product();
+    }
+
     public Product saveProduct(Product product) {
-        return productRepository.save(product);
+        return productRepository.existsProductByProductName(product.getProductName()) ?
+                productRepository.findProductByProductName(product.getProductName()) :
+                productRepository.save(product);
     }
 
     public void deleteById(Integer id) {
+        if (productRepository.existsById(id) == false)
+            throw new RuntimeException("No product with such id " + id);
         productRepository.deleteById(id);
     }
